Sheet 1-G_Summation_from_1_to_N

+++ Assuit ICPC/Sheet #1/G_Summation_from_1_to_N.cpp
@@ -3,9 +3,9 @@ using namespace std;
 
 int main() {
       
-    int N;
+    long long N;
     cin >> N;
-    int sum =N*(N+1)/2;
+    long long int sum =N*(N+1)/2;
     cout << sum << endl;
     return 0;
 }
